Convert API tests to new test fixtures
Affects: https://www.notion.so/lyrasis/Refactor-API-test-suite-9852f7a8e4484cc6bed8235acb20a7fe
