+++ pyproject.toml
@@ -56,6 +56,7 @@ files = [
     "customlists",
     "core",
     "tests/core",
+    "tests/api_new"
 ]
 plugins = ["sqlmypy"]
 warn_redundant_casts = true
