+++ tests/api_new/conftest.py
@@ -1,0 +1,13 @@
+pytest_plugins = [
+    "tests.fixtures.csv_files",
+    "tests.fixtures.database",
+    "tests.fixtures.marc_files",
+    "tests.fixtures.opds2_files",
+    "tests.fixtures.opds_files",
+    "tests.fixtures.overdrive",
+    "tests.fixtures.s3",
+    "tests.fixtures.sample_covers",
+    "tests.fixtures.search",
+    "tests.fixtures.time",
+    "tests.fixtures.vendor_id",
+]
